Clean up and split TensorFlow deps of text.py
